Add once_cell dependency and refactor message processing in daemon with peel_forward module
