+++ Cargo.toml
@@ -56,6 +56,7 @@ base64 = "0.21.5"
 itertools = "0.11.0"
 moro = "0.4.0"
 replay_filter = "0.1.2"
+once_cell = "1.18.0"
 
 [profile.dev]
 panic = 'abort'
